claude-telegram/requirements.txt
fastapi
uvicorn
httpx
google-api-python-client
google-auth
google-auth-httplib2
mcp[cli]
